Update 9th Gate for Sinar Pure Foods International Preparation for Go Live Finished Goods Modules 6th Commit for Finished Goods Modules
Updates:
+ Purchase
	+ Purchase Requisition
		/ Handle item by typing, not select item from list down

+ Non Fish
	+ Stores Withdrawal
		/ Change data type sws_qty from int to numeric(18, 2)
	
+ Finished Goods
	+ Reclassification
		+ Create document function
		+ List function
	+ Order Fill Authority to Label
		/ PTS get data using item id and pts number to prevent miss reclassification value
	+ Transfer
		/ PTS list, using loop instead get first row
	+ Delivery
		+ Create document function
		+ List function
	+ Return Cargo
		+ Create document function
		+ List function

+++ src/java/com/app/wms/engine/db/dao/spring/SwsDtlDaoImpl.java
@@ -31,7 +31,7 @@ public class SwsDtlDaoImpl extends AbstractDAO
         sd.setId(rs.getInt("id"));
         sd.setSwsCode(rs.getString("sws_code"));
         sd.setProductCode(rs.getString("product_code"));
-        sd.setQty(rs.getInt("qty"));
+        sd.setQty(rs.getBigDecimal("qty"));
         sd.setSoh(rs.getInt("soh"));
         sd.setUom(rs.getString("uom"));
         sd.setCreatedBy(rs.getString("created_by"));
